+++ src/Compute/Compute.Test/ScenarioTests/SqlIaaSExtensionTests.cs
@@ -30,12 +30,8 @@ namespace Microsoft.Azure.Commands.Compute.Test.ScenarioTests
             TestRunner.RunTestScript("Test-SetAzureRmVMSqlServerExtension");
         }
 
-#if NETSTANDARD
         [Fact(Skip = "Resources -> ResourceManager, needs re-recorded")]
         [Trait(Category.RunType, Category.DesktopOnly)]
-#else
-        [Fact(Skip ="CRP needs to re-record the test")]
-#endif
         [Trait(Category.AcceptanceType, Category.CheckIn)]
         public void TestSqlIaaSAKVExtension()
         {