+++ src/DataFactory/DataFactoryV2/Changelog.md
@@ -19,6 +19,7 @@
 -->
 
 ## Upcoming Release
+* Updating the output of get activity runs, get pipeline runs, and get trigger runs ADF cmdlets to support Select-Object pipe.
 
 ## Version 1.1.1
 * Add SsisProperties if NodeCount not null for managed integration runtime.